+++ crates/api_models/src/enums.rs
@@ -94,7 +94,7 @@ pub enum Connector {
     Cryptopay,
     Cybersource,
     Datatrans,
-    // Deutschebank,
+    Deutschebank,
     Dlocal,
     Ebanx,
     Fiserv,
@@ -187,6 +187,7 @@ impl Connector {
         matches!(
             (self, payment_method),
             (Self::Airwallex, _)
+                | (Self::Deutschebank, _)
                 | (Self::Globalpay, _)
                 | (Self::Paypal, _)
                 | (Self::Payu, _)
@@ -231,7 +232,7 @@ impl Connector {
             | Self::Cashtocode
             | Self::Coinbase
             | Self::Cryptopay
-			// | Self::Deutschebank
+			| Self::Deutschebank
             | Self::Dlocal
             | Self::Ebanx
             | Self::Fiserv
@@ -506,6 +507,7 @@ pub enum FieldType {
     UserShippingAddressCountry { options: Vec<String> },
     UserBlikCode,
     UserBank,
+    UserBankAccountNumber,
     Text,
     DropDown { options: Vec<String> },
     UserDateOfBirth,