+++ crates/hyperswitch_connectors/src/default_implementations.rs
@@ -100,8 +100,10 @@ default_imp_for_authorize_session_token!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
+    connectors::Volt,
     connectors::Thunes,
     connectors::Tsys,
     connectors::Worldline
@@ -129,6 +131,7 @@ default_imp_for_calculate_tax!(
     connectors::Helcim,
     connectors::Stax,
     connectors::Novalnet,
+    connectors::Mollie,
     connectors::Nexixpay,
     connectors::Fiuu,
     connectors::Globepay,
@@ -136,6 +139,7 @@ default_imp_for_calculate_tax!(
     connectors::Powertranz,
     connectors::Thunes,
     connectors::Tsys,
+    connectors::Volt,
     connectors::Deutschebank
 );
 
@@ -161,6 +165,7 @@ default_imp_for_session_update!(
     connectors::Helcim,
     connectors::Stax,
     connectors::Taxjar,
+    connectors::Mollie,
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Fiuu,
@@ -169,7 +174,8 @@ default_imp_for_session_update!(
     connectors::Powertranz,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Deutschebank
+    connectors::Deutschebank,
+    connectors::Volt
 );
 
 use crate::connectors;
@@ -190,7 +196,6 @@ macro_rules! default_imp_for_complete_authorize {
 
 default_imp_for_complete_authorize!(
     connectors::Bitpay,
-    connectors::Deutschebank,
     connectors::Fiserv,
     connectors::Fiservemea,
     connectors::Fiuu,
@@ -202,7 +207,8 @@ default_imp_for_complete_authorize!(
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_incremental_authorization {
@@ -232,11 +238,13 @@ default_imp_for_incremental_authorization!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_create_customer {
@@ -263,13 +271,15 @@ default_imp_for_create_customer!(
     connectors::Fiuu,
     connectors::Globepay,
     connectors::Helcim,
+    connectors::Mollie,
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_connector_redirect_response {
@@ -303,7 +313,8 @@ default_imp_for_connector_redirect_response!(
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_pre_processing_steps{
@@ -333,11 +344,13 @@ default_imp_for_pre_processing_steps!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_post_processing_steps{
@@ -367,11 +380,13 @@ default_imp_for_post_processing_steps!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_approve {
@@ -401,11 +416,13 @@ default_imp_for_approve!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_reject {
@@ -435,11 +452,13 @@ default_imp_for_reject!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_webhook_source_verification {
@@ -469,11 +488,13 @@ default_imp_for_webhook_source_verification!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_accept_dispute {
@@ -504,11 +525,13 @@ default_imp_for_accept_dispute!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_submit_evidence {
@@ -538,11 +561,13 @@ default_imp_for_submit_evidence!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_defend_dispute {
@@ -572,11 +597,13 @@ default_imp_for_defend_dispute!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_file_upload {
@@ -615,11 +642,13 @@ default_imp_for_file_upload!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -651,11 +680,13 @@ default_imp_for_payouts_create!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -687,11 +718,13 @@ default_imp_for_payouts_retrieve!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -723,11 +756,13 @@ default_imp_for_payouts_eligibility!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -759,11 +794,13 @@ default_imp_for_payouts_fulfill!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -795,11 +832,13 @@ default_imp_for_payouts_cancel!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -831,11 +870,13 @@ default_imp_for_payouts_quote!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -867,11 +908,13 @@ default_imp_for_payouts_recipient!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "payouts")]
@@ -903,11 +946,13 @@ default_imp_for_payouts_recipient_account!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "frm")]
@@ -939,11 +984,13 @@ default_imp_for_frm_sale!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "frm")]
@@ -975,11 +1022,13 @@ default_imp_for_frm_checkout!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "frm")]
@@ -1011,11 +1060,13 @@ default_imp_for_frm_transaction!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "frm")]
@@ -1047,11 +1098,13 @@ default_imp_for_frm_fulfillment!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 #[cfg(feature = "frm")]
@@ -1083,11 +1136,13 @@ default_imp_for_frm_record_return!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );
 
 macro_rules! default_imp_for_revoking_mandates {
@@ -1116,9 +1171,11 @@ default_imp_for_revoking_mandates!(
     connectors::Novalnet,
     connectors::Nexixpay,
     connectors::Powertranz,
+    connectors::Mollie,
     connectors::Stax,
     connectors::Taxjar,
     connectors::Thunes,
     connectors::Tsys,
-    connectors::Worldline
+    connectors::Worldline,
+    connectors::Volt
 );